requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Naked==0.1.32
packaging==25.0
pycryptodome==3.22.0
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
urllib3==2.4.0
Werkzeug==3.1.3
